+++ HISTORY.rst
@@ -1,6 +1,13 @@
 History
 -------
 
+0.0.2
++++++
+
+Added the Path class - string-like with shortcuts
+cd doubles as a function and a contextmanager
+Many extra aliases in ffs.nix
+
 0.0.1 (Jul 20 1012)
 +++++++++++++++++++